Added mailling system in the project

+++ ActPro/Areas/Admin/Controllers/ReservationsController.cs
@@ -1,12 +1,13 @@
 ﻿using ActPro.Services.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace ActPro.Areas.Admin.Controllers
 {
     [Area("Admin")]
     [Authorize(Roles = "Admin")]
-    public class ReservationsController(IReservationDashboardService resService) : Controller
+    public class ReservationsController(IReservationDashboardService resService, DAL.Data.ApplicationDbContext context, Services.Interfaces.IEmailSender emailSender) : Controller
     {
         //--- MENU ---
         [HttpGet]
@@ -20,9 +21,25 @@ namespace ActPro.Areas.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> Delete(int id)
         {
-            if (await resService.DeleteReservationAsync(id))
-                TempData["Success"] = "Резервацията беше анулирана.";
+            var res = await context.Reservations
+            .Include(r => r.AspNetUser)
+            .Include(r => r.Place)
+            .FirstOrDefaultAsync(r => r.Id == id);
 
+            if (res != null)
+            {
+                string userEmail = res.AspNetUser.Email;
+                string userFirstName = res.AspNetUser.FirstName;
+                string placeName = res.Place.Name;
+                string formattedDate = res.ReservationDate.ToString();
+                string timeSlot = res.ReservationTime.ToString();
+
+                if (await resService.DeleteReservationAsync(id))
+                {
+                    await emailSender.SendBookingCancellationAsync(userEmail, userFirstName, placeName, formattedDate, timeSlot);
+                    TempData["Success"] = "Резервацията беше анулирана";
+                }
+            }
             return RedirectToAction(nameof(Index));
         }
 
@@ -31,8 +48,27 @@ namespace ActPro.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> EditTime(int id, TimeOnly reservationTime)
         {
+            var res = await context.Reservations
+            .Include(r => r.AspNetUser)
+            .Include(r => r.Place)
+            .FirstOrDefaultAsync(r => r.Id == id);
+
+            if (res == null) return NotFound();
+
+            string oldTime = res.ReservationTime.ToString();
+            string newTime = reservationTime.ToString("HH:mm");
+
             if (await resService.UpdateReservationTimeAsync(id, reservationTime))
+            {
+                await emailSender.SendReservationTimeChangedAsync(
+                    res.AspNetUser.Email,
+                    res.AspNetUser.FirstName,
+                    res.Place.Name,
+                    oldTime,
+                    newTime,
+                    res.ReservationDate.ToString());
                 TempData["Success"] = "Часът беше променен успешно!";
+            }
 
             return RedirectToAction(nameof(Index));
         }
